rename the Contacts views
