+++ source/views/contacts/detail.js
@@ -1,7 +1,8 @@
 // @flow
+
 import * as React from 'react'
 import {Alert, StyleSheet, ScrollView, Image} from 'react-native'
-import {contactImages} from '../../../images/contact-images'
+import {contactImages} from '../../../images/contacts'
 import {Markdown} from '../components/markdown'
 import {ListFooter} from '../components/list'
 import glamorous from 'glamorous-native'
