+++ source/views/contacts/list.js
@@ -4,7 +4,7 @@ import * as React from 'react'
 import {SectionList, StyleSheet} from 'react-native'
 import {ListSeparator, ListSectionHeader} from '../components/list'
 import {ListEmpty} from '../components/list'
-import {ContactRow} from './contact-row'
+import {ContactRow} from './row'
 import delay from 'delay'
 import {reportNetworkProblem} from '../../lib/report-network-problem'
 import * as defaultData from '../../../docs/contact-info.json'
